chore: Removed the typedef patch (for autocxx)

+++ examples/inspect.rs
@@ -49,17 +49,10 @@ pub fn main() -> anyhow::Result<()> {
             println!("{}: {:?}", x.name(), tensor.shape());
             tensor.fill(1.0f32);
         });
-        // time!(interpreter.run_session_with_callback(&session, |_, name| {
-        //     println!("Before Callback: {:?}", name);
-        //     true
-        // },|_ , name| {
-        //     println!("After Callback: {:?}", name);
-        //     true
-        // } , true)?;"run session");
         interpreter.run_session(&session)?;
         let outputs = interpreter.outputs(&session);
         outputs.iter().for_each(|x| {
-            let tensor = x.tensor::<f32>().expect("No tensor");
+            let tensor = x.tensor::<u8>().expect("No tensor");
             time!(tensor.wait(ffi::MapType::MAP_TENSOR_READ, true); format!("Waiting for tensor: {}", x.name()));
             println!("{}: {:?}", x.name(), tensor.shape());
             let _ = tensor.create_host_tensor_from_device(true);
